Rename Equations.* in Valencia Sources.*
