+++ src/Evolution/Systems/RelativisticEuler/Valencia/Sources.cpp
@@ -1,7 +1,7 @@
 // Distributed under the MIT License.
 // See LICENSE.txt for details.
 
-#include "Evolution/Systems/RelativisticEuler/Valencia/Equations.hpp"
+#include "Evolution/Systems/RelativisticEuler/Valencia/Sources.hpp"
 
 #include "DataStructures/DataVector.hpp"
 #include "DataStructures/Tensor/Tensor.hpp"
