+++ tests/Unit/Evolution/Systems/RelativisticEuler/Valencia/CMakeLists.txt
@@ -6,9 +6,9 @@ set(LIBRARY "Test_Valencia")
 set(LIBRARY_SOURCES
   Test_Characteristics.cpp
   Test_ConservativeFromPrimitive.cpp
-  Test_Equations.cpp
   Test_Fluxes.cpp
   Test_PrimitiveFromConservative.cpp
+  Test_Sources.cpp
   Test_Tags.cpp
   )
